Got rid of all the .part calls, CodeGrid access is transparent now
Any indexing on the outer array returns a CodeGridRow.

Any indexing on a CodeGridRow returns a CodeGridPosition.  Any
assignment to a CodeGridRow accepts either a CodeGridPosition or Part
as argument, and will call .part on it, retrieving and assigning the
actual Part.

Any == (or != by extension) sent to a CodeGridPosition will use the
referenced Part instead.  Any messages not understood by a
CodeGridPosition will be forwarded to the referenced part.